% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{analyzeDesign}
\alias{analyzeDesign}
\title{Analyze (using ANOVA) a design using center points or some effects to estimate standard error}
\usage{
analyzeDesign(dm, obsCol = "obs", centerpoints = NULL,
  errorEst = NULL)
}
\arguments{
\item{dm}{A design matrix.  Generate these with designMatrix() and add observations.}

\item{obsCol}{The name of the column with observations in it.}

\item{centerpoints}{A vector of centerpoints used to estimate standard error (either this or errorEst are required)}

\item{errorEst}{A vector of effects to use to estimate standard error (either this or centerpoints are required)}
}
\value{
A dataframe with a columns for factor, sum of squares, degrees of freedom, mean squares, F-statistics, and P-values
}
\description{
This function estimates the effects of factors in an experiment (single replication...for now) and uses either centerpoints
or pools some effects to estimate standard error.  This allows for the calculation of F statistics and P values in the ANOVA analysis.
}
\examples{
##Analysis using high-order interactions to estimate error:

#create a design matrix with observations
z <- designMatrix(c('A','B','C','D'), standardOrder = TRUE)
z <- cbind(z, obs = c(550,669,604,650,633,642,601,635,1037,749,1052,868,1075,860, 1063, 729))

#analyze the design using 3 and 4-factor interactions to estimate error
analyzeDesign(z, errorEst = c('ABC','ABD','ACD','BCD','ABCD'))


##Analysis using centerpoints to estimate error:

#create a design matrix with observations
z <- designMatrix(c('A','B'), standardOrder = TRUE)
z <- cbind(z, obs = c(39.3,40.9,40,41.5))

#analyze the design using centerpoints to estimate error
analyzeDesign(z, errorEst = c('ABC','ABD','ACD','BCD','ABCD'))

}
