% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{designMatrix}
\alias{designMatrix}
\title{Create Design Matrix}
\usage{
designMatrix(factors, mainOnly = FALSE, standardOrder = TRUE)
}
\arguments{
\item{factors}{A vector of factors used in your experimental design - usually single letters.}

\item{mainOnly}{If set to TRUE, only columns with main effects (i.e., no interaction effects) will be created.}

\item{standardOrder}{If set to TRUE, effect columns will appear in standard order.  Otherwise, they will be reported in order of interaction.  No effect if mainOnly = TRUE.}
}
\value{
A matrix with a row for each treatment combination and a column for each main effect (and interaction effects, if included).
}
\description{
This function creates a design matrix for a given set of factors.  By convention, your factors should be single letters.
The user can determine if they want just the main effects, or all (interaction) effects included.
makeTreatments() is used to create the treatment column in the standard order and notation.
}
\examples{
##create full design matrix
designMatrix(factors = c('A','B','C','D'))

##create design matrix with main effects only
designMatrix(factors = c('A','B','C','D'), mainOnly = TRUE)

}
