% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_dotplot.R
\name{effect_dotplot}
\alias{effect_dotplot}
\title{Create a dot plot of effects.}
\usage{
effect_dotplot(model, scaling = 1, limits = NULL, ...)
}
\arguments{
\item{model}{Linear model.}

\item{scaling}{Multiply all model coefficients by this factor.}

\item{limits}{Numeric vector \code{c(lower, upper)} Effects outsize this range are moved to the range boundaries.}

\item{...}{Optional arguments for \code{ggplot2::geom_dotplot}.}
}
\description{
Create a dot plot of effects.
}
\author{
Paul Jensen, \email{software@jensenlab.net}
}
