% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_dotplot.R
\name{effects}
\alias{effects}
\title{Get effects from a linear model.}
\usage{
effects(model, scaling = 1, no_scale = "(Intercept)", intercept = FALSE)
}
\arguments{
\item{model}{A linear model object (`lm` or `rsm`).}

\item{scaling}{Multiply all model coefficients by this factor.}

\item{no_scale}{Character vector of effects to not scale (default=`"(Intercept)"`).}

\item{intercept}{When `FALSE` (default), exclude the intercept.}
}
\description{
Get effects from a linear model.
}
\author{
Paul Jensen, \email{software@jensenlab.net}
}
