% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{estEffects}
\alias{estEffects}
\title{Estimate effects}
\usage{
estEffects(dm, obsCol = "obs")
}
\arguments{
\item{dm}{A design matrix.  Generate these with designMatrix() and add observations.}

\item{obsCol}{The name of the column with observations in it.}
}
\value{
A dataframe with a columns for factor and effect.
}
\description{
This function estimates the effects of factors in an experiment (single replication...for now).
Note, this function assumes you are using single letters to detnote main effects (and uses the number of main effects for computation).
}
\examples{
##Estimate the effect of all factors:

#create a design matrix
z <- designMatrix(c('A','B','C','D'), standardOrder = TRUE)

#add observations
z <- cbind(z, obs = c(550,669,604,650,633,642,601,635,1037,749,1052,868,1075,860, 1063, 729))

#estimate effects
estEffects(dm = z, obsCol = 'obs')


}
