% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farplot.R
\name{farplot}
\alias{farplot}
\title{Plot factors and a response.}
\usage{
farplot(
  data,
  response = NULL,
  factors = NULL,
  size = 1,
  pch = 19,
  factor_colors = c("red", "black"),
  normalize = "all",
  size_transform = sqrt,
  extension = 0.05
)
}
\arguments{
\item{data}{Data frame.}

\item{response}{Name of response column. `NULL` uses the last column.}

\item{factors}{Names of factor columns. `NULL` selects all columns except `response`.}

\item{size}{Default size, in `char`, or the response symbols. Reduce this if symbols overlap.}

\item{pch}{Symbol type for response.}

\item{factor_colors}{Colors for negative and positive factors.}

\item{normalize}{If `"all"` (default), factor sizes are normalized to the largest single factor code.
Otherwise, each row is normalized separately.}

\item{extension}{Axis extension factor; see `dataViewport`.}

\item{size_tranform}{Function applied to convert factor codes to symbol sizes. Default is `sqrt`.
`NULL` applies no transform.}
}
\description{
Plot factors and a response.
}
