% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_dotplot.R
\name{get_effects}
\alias{get_effects}
\title{Get effects from a linear model.}
\usage{
get_effects(model, scaling = 1, no_scale = "(Intercept)", intercept = FALSE)
}
\arguments{
\item{model}{A linear model object (\code{lm} or \code{rsm}).}

\item{scaling}{Multiply all model coefficients by this factor.}

\item{no_scale}{Character vector of effects to not scale (default=\code{"(Intercept)"}).}

\item{intercept}{When \code{FALSE} (default), exclude the intercept.}
}
\description{
Get effects from a linear model.
}
\author{
Paul Jensen, \email{software@jensenlab.net}
}
