% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{makeInteractionGrid}
\alias{makeInteractionGrid}
\title{Create all interact effects}
\usage{
makeInteractionGrid(myData, standardOrder = TRUE)
}
\arguments{
\item{myData}{A design matrix with main effects only.}

\item{standardOrder}{Passed from designMatrix.}
}
\value{
A matrix with a row for each treatment combination and a column for each main effect (and interaction effects, if included).
}
\description{
This is an internal function that is called by designMatrix when mainOnly = FALSE.
}
