% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreatments.R
\name{makeTreatments}
\alias{makeTreatments}
\title{Create Treatment Combinations in Standard Notation}
\usage{
makeTreatments(factors, out = NULL)
}
\arguments{
\item{factors}{A vector of factors used in your experimental design - usually single capital letters.}

\item{out}{Only used for recursion - leave this set to NULL.}
}
\value{
A vector of treatment combinations in standard notation (and order)
}
\description{
This function helps define the treatment combinations in the correct (standard) notation and order.
According to standard notation, the combinations should be lowercase letters, so all inputs will be transformed to lower case.
This function can be used alone, but is more likely just a side effect of calling designMatrix().
}
\examples{
##create treatment combinations
makeTreatments(factors = c('A','B','C','D'))

}
