% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{show_effects}
\alias{show_effects}
\alias{show_effects.default}
\alias{show_effects.lm}
\title{Show effect sizes.}
\usage{
show_effects(...)

\method{show_effects}{default}(
  coefs,
  scaling = 1,
  n = length(coefs),
  no_scale = "(Intercept)",
  ordered = "none",
  colsep = "   ",
  aligner = align_numbers,
  ...
)

\method{show_effects}{lm}(model, intercept = TRUE, ...)
}
\arguments{
\item{coefs}{A named numeric vector of coefficients from a `lm` model.}

\item{scaling}{Scale all non-intercept coefficients by this factor (default=1).}

\item{n}{Show up to `n` effects.}

\item{no_scale}{Character vector of effects to not scale (default=`"(Intercept)"`).}

\item{ordered}{By default (`ordered="none"`), the effects are ordered as they appear in the input.
Setting `order="abs"` orders by absolute value (magnitude), and `order="signed"` shows in
numerical order (largest to smallest, respecting sign).}

\item{colsep}{Separator between effect names and value (default=`"   "`).}

\item{model}{A linear model produced by `lm`.}

\item{intercept}{When `TRUE` (default), show the intercept as an effect.}

\item{integer_digits}{Max number of digits to display left of the decimal point. Larger numbers
are displayed using scientific notation.}

\item{decimal_digits}{Effects are rounded to this many digits, and training zeros are not shown.}
}
\description{
Prints the effect sizes for a vector of coefficients from a linear model.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Base method for named numeric vectors

\item \code{lm}: Show effects from a linear model.
}}

\author{
Paul Jensen, \email{software@jensenlab.net}
}
